stellar_base
Module crypto

Cryptographic functions.

For long running programs, you should initialize the sodium library for optimal performance.

use stellar_base::crypto;

crypto::init()?;
// your program here.

Structs

KeyPair

The secret and public key pair of the account.

MuxedEd25519PublicKey

A public key together with an id.

PublicKey

The public key of the account.

SecretKey

The secret key of the account.

Enums

MuxedAccount

Functions

decode_account_id

decode_muxed_account

decode_pre_auth_tx

decode_secret_seed

decode_sha256_hash

encode_account_id

encode_muxed_account

encode_pre_auth_tx

encode_secret_seed

encode_sha256_hash

hash

Compute sha256 hash of m.

init

Initialize the sodium library and chooses faster version of the primitives if possible.

random_bytes

Generate size random bytes.