stellar_base::operations
Enum Operation

pub enum Operation {

    CreateAccount(CreateAccountOperation),
    Payment(PaymentOperation),
    PathPaymentStrictReceive(PathPaymentStrictReceiveOperation),
    ManageSellOffer(ManageSellOfferOperation),
    CreatePassiveSellOffer(CreatePassiveSellOfferOperation),
    SetOptions(SetOptionsOperation),
    ChangeTrust(ChangeTrustOperation),
    AllowTrust(AllowTrustOperation),
    AccountMerge(AccountMergeOperation),
    Inflation(InflationOperation),
    ManageData(ManageDataOperation),
    BumpSequence(BumpSequenceOperation),
    ManageBuyOffer(ManageBuyOfferOperation),
    PathPaymentStrictSend(PathPaymentStrictSendOperation),
    CreateClaimableBalance(CreateClaimableBalanceOperation),
    ClaimClaimableBalance(ClaimClaimableBalanceOperation),
    BeginSponsoringFutureReserves(BeginSponsoringFutureReservesOperation),
    EndSponsoringFutureReserves(EndSponsoringFutureReservesOperation),
    RevokeSponsorship(RevokeSponsorshipOperation),

}

Operations on a Stellar network.

Variants

CreateAccount(CreateAccountOperation)

Create and fund a non existing account.

Payment(PaymentOperation)

Create a payment.

PathPaymentStrictReceive(PathPaymentStrictReceiveOperation)

Send the specified amount of asset, optionally through a path.

ManageSellOffer(ManageSellOfferOperation)

Create, update, or delete a sell offer.

CreatePassiveSellOffer(CreatePassiveSellOfferOperation)

Create a passive sell offer. This offer won’t consume a counter offer that exactly matches this offer.

SetOptions(SetOptionsOperation)

Change the account options.

ChangeTrust(ChangeTrustOperation)

Add, remove, or update a trust line for the given asset.

AllowTrust(AllowTrustOperation)

Authorize another account to hold your account’s credit aset.

AccountMerge(AccountMergeOperation)

Transfer all of account balance into destination account.

Inflation(InflationOperation)

Generate the inflation.

ManageData(ManageDataOperation)

Add o remove data entry.

BumpSequence(BumpSequenceOperation)

Bumps the account sequence number.

ManageBuyOffer(ManageBuyOfferOperation)

Create, update, or delete a buy offer.

PathPaymentStrictSend(PathPaymentStrictSendOperation)

Send the specified amount of asset, optionally through a path.

CreateClaimableBalance(CreateClaimableBalanceOperation)

Create a new claimable balance.

ClaimClaimableBalance(ClaimClaimableBalanceOperation)

Claim a claimable balance.

BeginSponsoringFutureReserves(BeginSponsoringFutureReservesOperation)

Begin sponsoring future reserves for an account. Needs a matching EndSponsoringFutureReserves in the same transaction.

EndSponsoringFutureReserves(EndSponsoringFutureReservesOperation)

Ends a BeginSponsoringFutureReserves operation in the transaction.

RevokeSponsorship(RevokeSponsorshipOperation)

Revoke a reserve sponsorship.

Implementations

impl Operation

pub fn new_create_account() -> CreateAccountOperationBuilder

Creates a new create account operation builder.

pub fn new_payment() -> PaymentOperationBuilder

Creates a new payment operation builder.

pub fn new_path_payment_strict_receive() -> PathPaymentStrictReceiveOperationBuilder

Creates a new path payment strict receive operation builder.

pub fn new_manage_sell_offer() -> ManageSellOfferOperationBuilder

Creates a new manage sell offer operation builder.

pub fn new_create_passive_sell_offer() -> CreatePassiveSellOfferOperationBuilder

Creates a new create passive sell offer operation builder.

pub fn new_set_options() -> SetOptionsOperationBuilder

Creates a new set options operation builder.

pub fn new_change_trust() -> ChangeTrustOperationBuilder

Creates a new change trust operation builder.

pub fn new_allow_trust() -> AllowTrustOperationBuilder

Creates a new allow trust operation builder.

pub fn new_account_merge() -> AccountMergeOperationBuilder

Creates a new account merge operation builder.

pub fn new_inflation() -> InflationOperationBuilder

Creates a new inflation operation builder.

pub fn new_manage_data() -> ManageDataOperationBuilder

Creates a new manage data operation builder.

pub fn new_bump_sequence() -> BumpSequenceOperationBuilder

Creates a new bump sequence operation builder.

pub fn new_manage_buy_offer() -> ManageBuyOfferOperationBuilder

Creates a new manage buy offer operation builder.

pub fn new_path_payment_strict_send() -> PathPaymentStrictSendOperationBuilder

Creates a new path payment strict send operation builder.

pub fn new_create_claimable_balance() -> CreateClaimableBalanceOperationBuilder

Creates a new create claimable balance operation builder.

pub fn new_claim_claimable_balance() -> ClaimClaimableBalanceOperationBuilder

Creates a new claim claimable balance operation builder.

pub fn new_begin_sponsoring_future_reserves() -> BeginSponsoringFutureReservesOperationBuilder

Creates a new begin sponsoring future reserves operation builder.

pub fn new_end_sponsoring_future_reserves() -> EndSponsoringFutureReservesOperationBuilder

Creates a new end sponsoring future reserves operation builder.

pub fn new_revoke_sponsorship() -> RevokeSponsorshipOperationBuilder

Creates a new revoke sponsorship operation builder.

pub fn as_create_account(&self) -> Option<&CreateAccountOperation>

If the operation is a CreateAccount, returns its value. Returns None otherwise.

pub fn as_create_account_mut(&mut self) -> Option<&mut CreateAccountOperation>

If the operation is a CreateAccount, returns its mutable value. Returns None otherwise.

pub fn is_create_account(&self) -> bool

Returns true if the operation is a CreateAccount.

pub fn as_payment(&self) -> Option<&PaymentOperation>

If the operation is a Payment, returns its value. Returns None otherwise.

pub fn as_payment_mut(&mut self) -> Option<&mut PaymentOperation>

If the operation is a Payment, returns its value. Returns None otherwise.

pub fn is_payment(&self) -> bool

Returns true if the operation is a Payment.

pub fn as_path_payment_strict_receive( &self, ) -> Option<&PathPaymentStrictReceiveOperation>

If the operation is a PathPaymentStrictReceive, returns its value. Returns None otherwise.

pub fn as_path_payment_strict_receive_mut( &mut self, ) -> Option<&mut PathPaymentStrictReceiveOperation>

If the operation is a PathPaymentStrictReceive, returns its mutable value. Returns None otherwise.

pub fn is_path_payment_strict_receive(&self) -> bool

Returns true if the operation is a PathPaymentStrictReceive.

pub fn as_manage_sell_offer(&self) -> Option<&ManageSellOfferOperation>

If the operation is a ManageSellOffer, returns its value. Returns None otherwise.

pub fn as_manage_sell_offer_mut( &mut self, ) -> Option<&mut ManageSellOfferOperation>

If the operation is a ManageSellOffer, returns its mutable value. Returns None otherwise.

pub fn is_manage_sell_offer(&self) -> bool

Returns true if the operation is a ManageSellOffer.

pub fn as_create_passive_sell_offer( &self, ) -> Option<&CreatePassiveSellOfferOperation>

If the operation is a CreatePassiveSellOffer, returns its value. Returns None otherwise.

pub fn as_create_passive_sell_offer_mut( &mut self, ) -> Option<&mut CreatePassiveSellOfferOperation>

If the operation is a CreatePassiveSellOffer, returns its mutable value. Returns None otherwise.

pub fn is_create_passive_sell_offer(&self) -> bool

Returns true if the operation is a CreatePassiveSellOffer.

pub fn as_set_options(&self) -> Option<&SetOptionsOperation>

If the operation is a SetOptions, returns its value. Returns None otherwise.

pub fn as_set_options_mut(&mut self) -> Option<&mut SetOptionsOperation>

If the operation is a SetOptions, returns its mutable value. Returns None otherwise.

pub fn is_set_options(&self) -> bool

Returns true if the operation is a SetOptions.

pub fn as_change_trust(&self) -> Option<&ChangeTrustOperation>

If the operation is a ChangeTrust, returns its value. Returns None otherwise.

pub fn as_change_trust_mut(&mut self) -> Option<&ChangeTrustOperation>

If the operation is a ChangeTrust, returns its mutable value. Returns None otherwise.

pub fn is_change_trust(&self) -> bool

Returns true if the operation is a ChangeTrust.

pub fn as_allow_trust(&self) -> Option<&AllowTrustOperation>

If the operation is a AllowTrust, returns its value. Returns None otherwise.

pub fn as_allow_trust_mut(&mut self) -> Option<&mut AllowTrustOperation>

If the operation is a AllowTrust, returns its mutable value. Returns None otherwise.

pub fn is_allow_trust(&self) -> bool

Returns true if the operation is a AllowTrust.

pub fn as_account_merge(&self) -> Option<&AccountMergeOperation>

If the operation is a AccountMerge, returns its value. Returns None otherwise.

pub fn as_account_merge_mut(&mut self) -> Option<&mut AccountMergeOperation>

If the operation is a AccountMerge, returns its mutable value. Returns None otherwise.

pub fn is_account_merge(&self) -> bool

Returns true if the operation is a AccountMerge.

pub fn as_inflation(&self) -> Option<&InflationOperation>

If the operation is a Inflation, returns its value. Returns None otherwise.

pub fn as_inflation_mut(&mut self) -> Option<&mut InflationOperation>

If the operation is a Inflation, returns its mutable value. Returns None otherwise.

pub fn is_inflation(&self) -> bool

Returns true if the operation is a Inflation.

pub fn as_manage_data(&self) -> Option<&ManageDataOperation>

If the operation is a ManageData, returns its value. Returns None otherwise.

pub fn as_manage_data_mut(&mut self) -> Option<&mut ManageDataOperation>

If the operation is a ManageData, returns its mutable value. Returns None otherwise.

pub fn is_manage_data(&self) -> bool

Returns true if the operation is a ManageData.

pub fn as_bump_sequence(&self) -> Option<&BumpSequenceOperation>

If the operation is a BumpSequence, returns its value. Returns None otherwise.

pub fn as_bump_sequence_mut(&mut self) -> Option<&mut BumpSequenceOperation>

If the operation is a BumpSequence, returns its mutable value. Returns None otherwise.

pub fn is_bump_sequence(&self) -> bool

Returns true if the operation is a BumpSequence.

pub fn as_manage_buy_offer(&self) -> Option<&ManageBuyOfferOperation>

If the operation is a ManageBuyOffer, returns its value. Returns None otherwise.

pub fn as_manage_buy_offer_mut( &mut self, ) -> Option<&mut ManageBuyOfferOperation>

If the operation is a ManageBuyOffer, returns its mutable value. Returns None otherwise.

pub fn is_manage_buy_offer(&self) -> bool

Returns true if the operation is a ManageBuyOffer.

pub fn as_path_payment_strict_send( &self, ) -> Option<&PathPaymentStrictSendOperation>

If the operation is a PathPaymentStrictSend, returns its value. Returns None otherwise.

pub fn as_path_payment_strict_send_mut( &mut self, ) -> Option<&mut PathPaymentStrictSendOperation>

If the operation is a PathPaymentStrictSend, returns its mutable value. Returns None otherwise.

pub fn is_path_payment_strict_send(&self) -> bool

Returns true if the operation is a PathPaymentStrictSend.

pub fn as_create_claimable_balance( &self, ) -> Option<&CreateClaimableBalanceOperation>

If the operation is a CreateClaimableBalance, returns its value. Returns None otherwise.

pub fn as_create_claimable_balance_mut( &mut self, ) -> Option<&mut CreateClaimableBalanceOperation>

If the operation is a CreateClaimableBalance, returns its mutable value. Returns None otherwise.

pub fn is_create_claimable_balance(&self) -> bool

Returns true if the operation is a CreateClaimableBalance.

pub fn as_claim_claimable_balance( &self, ) -> Option<&ClaimClaimableBalanceOperation>

If the operation is a ClaimClaimableBalance, returns its value. Returns None otherwise.

pub fn as_claim_claimable_balance_mut( &mut self, ) -> Option<&mut ClaimClaimableBalanceOperation>

If the operation is a ClaimClaimableBalance, returns its mutable value. Returns None otherwise.

pub fn is_claim_claimable_balance(&self) -> bool

Returns true if the operation is a ClaimClaimableBalance.

pub fn as_begin_sponsoring_future_reserves( &self, ) -> Option<&BeginSponsoringFutureReservesOperation>

If the operation is a BeginSponsoringFutureReserves, returns its value. Returns None otherwise.

pub fn as_begin_sponsoring_future_reserves_mut( &mut self, ) -> Option<&mut BeginSponsoringFutureReservesOperation>

If the operation is a BeginSponsoringFutureReserves, returns its mutable value. Returns None otherwise.

pub fn is_begin_sponsoring_future_reserves(&self) -> bool

Returns true if the operation is a BeginSponsoringFutureReserves.

pub fn as_end_sponsoring_future_reserves( &self, ) -> Option<&EndSponsoringFutureReservesOperation>

If the operation is a EndSponsoringFutureReserves, returns its value. Returns None otherwise.

pub fn as_end_sponsoring_future_reserves_mut( &mut self, ) -> Option<&mut EndSponsoringFutureReservesOperation>

If the operation is a EndSponsoringFutureReserves, returns its mutable value. Returns None otherwise.

pub fn is_end_sponsoring_future_reserves(&self) -> bool

Returns true if the operation is a EndSponsoringFutureReserves.

pub fn as_revoke_sponsorship(&self) -> Option<&RevokeSponsorshipOperation>

If the operation is a RevokeSponsorship, returns its value. Returns None otherwise.

pub fn as_revoke_sponsorship_mut( &mut self, ) -> Option<&mut RevokeSponsorshipOperation>

If the operation is a RevokeSponsorship, returns its mutable value. Returns None otherwise.

pub fn is_revoke_sponsorship(&self) -> bool

Returns true if the operation is a RevokeSponsorship.

pub fn source_account(&self) -> &Option<MuxedAccount>

Retrieves the operation source account.

pub fn source_account_mut(&mut self) -> &mut Option<MuxedAccount>

Retrieves a mutable reference to the operation source account.

pub fn to_xdr(&self) -> Result<Operation>

Returns the xdr object.

pub fn from_xdr(x: &Operation) -> Result<Operation>

Creates from the xdr object.