stellar_base::transaction_result
Enum InnerTransactionResult

pub enum InnerTransactionResult {

    Success(TransactionResultSuccess),
    Failed(TransactionResultFailed),
    TooEarly(TransactionResultTooEarly),
    TooLate(TransactionResultTooLate),
    MissingOperation(TransactionResultMissingOperation),
    BadSequence(TransactionResultBadSequence),
    BadAuth(TransactionResultBadAuth),
    InsufficientBalance(TransactionResultInsufficientBalance),
    NoAccount(TransactionResultNoAccount),
    InsufficientFee(TransactionResultInsufficientFee),
    BadAuthExtra(TransactionResultBadAuthExtra),
    InternalError(TransactionResultInternalError),
    NotSupported(TransactionResultNotSupported),
    BadSponsorship(TransactionResultBadSponsorship),

}

Result of the inner transaction in a FeeBumpTransaction.

Variants

Success(TransactionResultSuccess)

Failed(TransactionResultFailed)

TooEarly(TransactionResultTooEarly)

TooLate(TransactionResultTooLate)

MissingOperation(TransactionResultMissingOperation)

BadSequence(TransactionResultBadSequence)

BadAuth(TransactionResultBadAuth)

InsufficientBalance(TransactionResultInsufficientBalance)

NoAccount(TransactionResultNoAccount)

InsufficientFee(TransactionResultInsufficientFee)

BadAuthExtra(TransactionResultBadAuthExtra)

InternalError(TransactionResultInternalError)

NotSupported(TransactionResultNotSupported)

BadSponsorship(TransactionResultBadSponsorship)

Implementations

impl InnerTransactionResult

pub fn as_success(&self) -> Option<&TransactionResultSuccess>

If the result is a Success, returns its value. Returns None otherwise

pub fn is_success(&self) -> bool

Returns true if the transaction result is a Success.

pub fn as_failed(&self) -> Option<&TransactionResultFailed>

If the result is a Failed, returns its value. Returns None otherwise

pub fn is_failed(&self) -> bool

Returns true if the transaction result is a Failed.

pub fn as_too_early(&self) -> Option<&TransactionResultTooEarly>

If the result is a TooEarly, returns its value. Returns None otherwise

pub fn is_too_early(&self) -> bool

Returns true if the transaction result is a TooEarly.

pub fn as_too_late(&self) -> Option<&TransactionResultTooLate>

If the result is a TooLate, returns its value. Returns None otherwise

pub fn is_too_late(&self) -> bool

Returns true if the transaction result is a TooLate.

pub fn as_missing_operation(&self) -> Option<&TransactionResultMissingOperation>

If the result is a MissingOperation, returns its value. Returns None otherwise

pub fn is_missing_operation(&self) -> bool

Returns true if the transaction result is a MissingOperation.

pub fn as_bad_sequence(&self) -> Option<&TransactionResultBadSequence>

If the result is a BadSequence, returns its value. Returns None otherwise

pub fn is_bad_sequence(&self) -> bool

Returns true if the transaction result is a BadSequence.

pub fn as_bad_auth(&self) -> Option<&TransactionResultBadAuth>

If the result is a BadAuth, returns its value. Returns None otherwise

pub fn is_bad_auth(&self) -> bool

Returns true if the transaction result is a BadAuth.

pub fn as_insufficient_balance( &self, ) -> Option<&TransactionResultInsufficientBalance>

If the result is a InsufficientBalance, returns its value. Returns None otherwise

pub fn is_insufficient_balance(&self) -> bool

Returns true if the transaction result is a InsufficientBalance.

pub fn as_no_account(&self) -> Option<&TransactionResultNoAccount>

If the result is a NoAccount, returns its value. Returns None otherwise

pub fn is_no_account(&self) -> bool

Returns true if the transaction result is a NoAccount.

pub fn as_insufficient_fee(&self) -> Option<&TransactionResultInsufficientFee>

If the result is a InsufficientFee, returns its value. Returns None otherwise

pub fn is_insufficient_fee(&self) -> bool

Returns true if the transaction result is a InsufficientFee.

pub fn as_bad_auth_extra(&self) -> Option<&TransactionResultBadAuthExtra>

If the result is a BadAuthExtra, returns its value. Returns None otherwise

pub fn is_bad_auth_extra(&self) -> bool

Returns true if the transaction result is a BadAuthExtra.

pub fn as_internal_error(&self) -> Option<&TransactionResultInternalError>

If the result is a InternalError, returns its value. Returns None otherwise

pub fn is_internal_error(&self) -> bool

Returns true if the transaction result is a InternalError.

pub fn as_not_supported(&self) -> Option<&TransactionResultNotSupported>

If the result is a NotSupported, returns its value. Returns None otherwise

pub fn is_not_supported(&self) -> bool

Returns true if the transaction result is a NotSupported.

pub fn as_bad_sponsorship(&self) -> Option<&TransactionResultBadSponsorship>

If the result is a BadSponsorship, returns its value. Returns None otherwise

pub fn is_bad_sponsorship(&self) -> bool

Returns true if the transaction result is a BadSponsorship.

pub fn from_xdr(x: &InnerTransactionResult) -> Result<InnerTransactionResult>

Creates TransactionResult from xdr object.